pyproject.toml
[project]
name = "django-mini-project"
version = "0.1.0"
description = "Django mini project with dev/prod split"
requires-python = ">=3.10"

# 런타임 의존성 (uv add 로 자동 관리)
dependencies = [
  "django>=5.1",
  "djangorestframework>=3.15",
  "django-environ>=0.11",
  "drf-spectacular>=0.27",
  "drf-spectacular-sidecar>=2025.10.1",
  "psycopg2-binary>=2.9.11",
]

[dependency-groups]
prod = [
  "djangorestframework-simplejwt>=5.5.1",
  "gunicorn>=21",
  "psycopg2-binary>=2.9",
]
dev = [
  "ruff>=0.6",
  "pytest>=8.0",
  "pytest-django>=4.8",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.dev"
python_files = ["test_*.py", "*_tests.py", "tests.py"]
addopts = "-ra -q"

[tool.ruff]
target-version = "py310"
line-length = 100
extend-exclude = ["migrations"]
# 기본 선택(PEP8/pyflakes/pycodestyle 등)
select = ["E", "F", "I", "UP", "B"]
ignore = [
  "E203",  # black 과 충돌 가능
]
# 자동 정렬(imports)
[tool.ruff.lint.isort]
known-first-party = ["apps", "config"]
combine-as-imports = true
